updated donation delete

+++ src/main/java/com/project/services/competition/CompetitionServiceImpl.java
@@ -7,12 +7,14 @@ import com.project.models.Competition;
 import com.project.models.Sponsor;
 import com.project.models.Sport;
 import com.project.repositories.CompetitionRepository;
+import com.project.repositories.DonationRepository;
 import com.project.services.sport.SportService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 import org.springframework.web.server.ResponseStatusException;
 
+import javax.ws.rs.BadRequestException;
 import java.util.List;
 import java.util.stream.Collectors;
 
@@ -23,11 +25,15 @@ public class CompetitionServiceImpl implements CompetitionService {
 
     private final SportService sportService;
 
+    private final DonationRepository donationRepository;
+
     @Autowired
     public CompetitionServiceImpl(CompetitionRepository competitionRepository,
-                                  SportService sportService) {
+                                  SportService sportService,
+                                  DonationRepository donationRepository) {
         this.competitionRepository = competitionRepository;
         this.sportService = sportService;
+        this.donationRepository = donationRepository;
     }
 
     @Override
@@ -85,6 +91,8 @@ public class CompetitionServiceImpl implements CompetitionService {
 
     /**
      * When deleting a competition, sponsors will not be deleted
+     * If a competition has already made donations, it can not be deleted, as the money
+     * have already been sent
      */
     @Override
     public void deleteCompetition(Long id) {
@@ -92,6 +100,10 @@ public class CompetitionServiceImpl implements CompetitionService {
         List<Sponsor> sponsors = competition.getSponsors();
         competition.getSponsors().removeAll(sponsors);
         competitionRepository.save(competition);
+
+        if (competition.getDonations().size() > 0) {
+            throw new BadRequestException("Competition can not be deleted as donations are in progress!");
+        }
 
         competitionRepository.delete(competition);
     }
